verify_login.py
import requests

url = 'http://localhost:5000/api/auth/login'
data = {'username': 'admin', 'password': 'adminpass'}

try:
    response = requests.post(url, json=data)
    print(f"Status: {response.status_code}")
    print(f"Response: {response.json()}")
except Exception as e:
    print(f"Error: {e}")
