requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Migrate
PyJWT
python-dotenv
email-validator
